pyproject.toml
[project]
name = "template-fastapi-next-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "uvicorn>=0.34.2",
    "python-dotenv>=1.1.0",
    "httpx>=0.25.0",
    "langchain>=1.2.13",
    "langchain-core>=1.2.23",
    "langchain-google-genai>=4.2.1",
    "langchain-google-vertexai>=3.2.2",
    "langgraph>=1.1.3",
    "langgraph-checkpoint>=4.0.1",
    "langgraph-prebuilt>=1.0.8",
    "langgraph-checkpoint-sqlite>=3.0.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
